requirements.txt
google_api_python_client == 2.51.0
google_auth == 2.8.0
google_auth_httplib2 == 0.1.0
google_auth_oauthlib == 0.5.2
httplib2 == 0.20.4
httpsproxy-urllib2 == 1.0
oauth2client == 4.1.3
oauthlib == 3.2.0
psycopg2_binary == 2.9.3
requests == 2.28.0
requests_oauthlib == 1.3.1

